+++ platforms/azure/giantnetes/variables.tf
@@ -128,7 +128,7 @@ variable "flatcar_linux_channel" {
 variable "flatcar_linux_version" {
   description = "Flatcar linux version."
   type        = string
-  default     = "2512.5.0"
+  default     = "2765.2.2"
 }
 
 variable "vault_image_publisher" {
@@ -152,7 +152,7 @@ variable "docker_registry" {
 }
 variable "hyperkube_version" {
   type    = string
-  default = "1.19.7"
+  default = "1.20.5"
 }
 
 ### DNS ###