+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Update Kubernetes to `1.20.5` version.
+- Update Kubernetes to `1.19.9` version.
 - Update Flatcar Linux to `2765.2.2` version in AWS.
 
 ## [3.4.0] - 2021-02-17